#6712 number format changed to double for CSV number converter (#6721)

+++ kernel/base/src/main/java/com/twosigma/beakerx/fileloader/CSV.java
@@ -61,7 +61,7 @@ public class CSV {
         return new BigInteger((String) value);
       } catch (Exception ignored) {
       }
-      return Float.parseFloat((String) value);
+      return Double.parseDouble((String) value);
     }
     return value;
   }
